v1.0.1

+++ src/lib/db.ts
@@ -14,7 +14,7 @@ export async function assertEvolutionsTable(prisma: PrismaClient, schema: string
     await prisma.$executeRawUnsafe(`CREATE SCHEMA IF NOT EXISTS ${schema};`);
 
     await prisma.$executeRawUnsafe(`CREATE TABLE IF NOT EXISTS ${schema}.${metaTable} (
-        id int not null primary key,
+        id bigint not null primary key,
         hash varchar(255) not null,
         applied_at timestamp not null,
         apply_script text,
